Java: Fix integer intervals
